+++ mqtt/mqtt-bridge/src/pump/mod.rs
@@ -1,22 +1,17 @@
-#![allow(dead_code, unused_imports, unused_variables)]
-use std::{
-    convert::TryInto,
-    fmt::{Display, Formatter, Result as FmtResult},
-    future::Future,
-    rc::Rc,
-};
+mod builder;
 
+pub use builder::Builder;
+
+// #![allow(dead_code, unused_imports, unused_variables)]
 use futures_util::{
-    future::{self, Either, FutureExt},
-    pin_mut,
     stream::{StreamExt, TryStreamExt},
+    FutureExt,
 };
 use tokio::{
     select,
     sync::{
         mpsc::{self, error::SendError},
         oneshot,
-        oneshot::Receiver,
     },
 };
 use tracing::{debug, error, info};
@@ -25,12 +20,11 @@ use mqtt3::PublishHandle;
 
 use crate::{
     bridge::BridgeError,
-    client::{ClientShutdownHandle, EventHandler, MqttClient, MqttClientConfig},
+    client::{ClientShutdownHandle, EventHandler, MqttClient},
     connectivity::ConnectivityState,
-    messages::{LocalUpstreamHandler, MessageHandler, TopicMapper},
     persist::{MessageLoader, PublicationStore, WakingMemoryStore},
-    rpc::{CommandId, LocalRpcHandler, RpcCommand},
-    settings::{ConnectionSettings, Credentials, TopicRule},
+    rpc::{CommandId, RpcCommand},
+    settings::ConnectionSettings,
 };
 
 #[derive(Debug, PartialEq)]
@@ -56,130 +50,6 @@ impl PumpHandle {
     }
 }
 
-pub struct Builder {
-    local: PumpBuilder,
-    remote: PumpBuilder,
-    store: Box<dyn Fn() -> PublicationStore<WakingMemoryStore>>,
-}
-
-impl Default for Builder {
-    fn default() -> Self {
-        Self {
-            local: PumpBuilder::default(),
-            remote: PumpBuilder::default(),
-            store: Box::new(|| PublicationStore::new_memory(0)),
-        }
-    }
-}
-
-impl Builder {
-    pub fn with_local<F>(&mut self, mut apply: F) -> &mut Self
-    where
-        F: FnMut(&mut PumpBuilder),
-    {
-        apply(&mut self.local);
-        self
-    }
-
-    pub fn with_remote<F>(&mut self, mut apply: F) -> &mut Self
-    where
-        F: FnMut(&mut PumpBuilder),
-    {
-        apply(&mut self.remote);
-        self
-    }
-
-    pub fn with_store<F>(&mut self, store: F) -> &mut Self
-    where
-        F: Fn() -> PublicationStore<WakingMemoryStore> + 'static,
-    {
-        self.store = Box::new(store);
-        self
-    }
-
-    pub fn build(
-        &mut self,
-    ) -> Result<
-        (
-            Pump<LocalUpstreamHandler<WakingMemoryStore>>,
-            Pump<MessageHandler<WakingMemoryStore>>,
-        ),
-        BridgeError,
-    > {
-        let ingress_store = (self.store)();
-        let ingress_loader = ingress_store.loader();
-
-        let egress_store = (self.store)();
-        let egress_loader = egress_store.loader();
-
-        let (remote_messages_send, remote_messages_recv) = mpsc::channel(100);
-        let (local_messages_send, local_messages_recv) = mpsc::channel(100);
-
-        let (subscriptions, topic_filters) = make_topics(&self.local.rules)?;
-
-        let rpc = LocalRpcHandler::new(PumpHandle::new(remote_messages_send.clone()));
-        let messages = MessageHandler::new(ingress_store.clone(), topic_filters);
-        let handler = LocalUpstreamHandler::new(messages, rpc);
-
-        let config = self.local.client.take().expect("local client config");
-        let client = MqttClient::tls(config, handler);
-        let local_pump = Pump::new(
-            local_messages_send,
-            local_messages_recv,
-            client,
-            subscriptions,
-            ingress_loader,
-            egress_store.clone(),
-        )?;
-
-        let (subscriptions, topic_filters) = make_topics(&self.remote.rules)?;
-
-        let handler = MessageHandler::new(egress_store, topic_filters);
-
-        let config = self.local.client.take().expect("local client config");
-        let client = MqttClient::tls(config, handler);
-        let remote_pump = Pump::new(
-            remote_messages_send,
-            remote_messages_recv,
-            client,
-            subscriptions,
-            egress_loader,
-            ingress_store,
-        )?;
-
-        Ok((local_pump, remote_pump))
-    }
-}
-
-#[derive(Default)]
-pub struct PumpBuilder {
-    // builder: Option<Box<Builder>>,
-    client: Option<MqttClientConfig>,
-    rules: Vec<TopicRule>,
-}
-
-impl PumpBuilder {
-    pub fn with_rules(&mut self, rules: &[TopicRule]) -> &mut Self {
-        self.rules = rules.to_vec();
-        self
-    }
-
-    pub fn with_config(&mut self, config: MqttClientConfig) -> &mut Self {
-        self.client = Some(config);
-        self
-    }
-}
-
-fn make_topics(rules: &[TopicRule]) -> Result<(Vec<String>, Vec<TopicMapper>), BridgeError> {
-    let (subscriptions, topic_rules): (Vec<_>, Vec<_>) = rules.iter().map(format_key_value).unzip();
-    let topic_filters = topic_rules
-        .into_iter()
-        .map(|topic| topic.try_into())
-        .collect::<Result<Vec<_>, _>>()?;
-
-    Ok((subscriptions, topic_filters))
-}
-
 #[cfg(test)]
 pub fn channel() -> (PumpHandle, mpsc::Receiver<PumpMessage>) {
     let (tx, rx) = tokio::sync::mpsc::channel(10);
@@ -189,75 +59,6 @@ pub fn channel() -> (PumpHandle, mpsc::Receiver<PumpMessage>) {
 #[derive(Debug, thiserror::Error)]
 #[error(transparent)]
 pub struct PumpError(SendError<PumpMessage>);
-
-pub fn local_pump(
-    connection_settings: &ConnectionSettings,
-    system_address: impl Into<String>,
-    device_id: impl AsRef<str>,
-    ingress_store: PublicationStore<WakingMemoryStore>,
-    egress_store: PublicationStore<WakingMemoryStore>,
-) -> Result<Pump<LocalUpstreamHandler<WakingMemoryStore>>, BridgeError> {
-    todo!()
-    // let client_id = format!(
-    //     "{}/$edgeHub/$bridge/{}",
-    //     device_id.as_ref(),
-    //     connection_settings.name()
-    // );
-
-    // let (subscriptions, topic_filters) = make_topics(connection_settings.forwards())?;
-
-    // let loader = ingress_store.loader();
-
-    // let messages = MessageHandler::new(ingress_store, topic_filters);
-
-    // let (sender, _) = tokio::sync::mpsc::channel(100); // TODO this should be remote handle
-    // let rpc = LocalRpcHandler::new(PumpHandle::new(sender));
-    // let handler = LocalUpstreamHandler::new(messages, rpc);
-
-    // let credentials = Credentials::Anonymous(client_id);
-
-    // let client = MqttClient::tls(
-    //     system_address,
-    //     connection_settings.keep_alive(),
-    //     connection_settings.clean_session(),
-    //     handler,
-    //     &credentials,
-    // );
-    // Ok(Pump::new(client, subscriptions, loader, egress_store)?)
-}
-
-pub fn remote_pump(
-    connection_settings: &ConnectionSettings,
-    ingress_store: PublicationStore<WakingMemoryStore>,
-    egress_store: PublicationStore<WakingMemoryStore>,
-) -> Result<Pump<MessageHandler<WakingMemoryStore>>, BridgeError> {
-    todo!()
-    // let (subscriptions, topic_filters) = make_topics(connection_settings.subscriptions())?;
-
-    // let loader = egress_store.loader();
-
-    // let handler = MessageHandler::new(egress_store, topic_filters);
-
-    // let credentials = connection_settings.credentials();
-
-    // let client = MqttClient::tls(
-    //     connection_settings.address(),
-    //     connection_settings.keep_alive(),
-    //     connection_settings.clean_session(),
-    //     handler,
-    //     &credentials,
-    // );
-    // Ok(Pump::new(client, subscriptions, loader, ingress_store)?)
-}
-
-fn format_key_value(topic: &TopicRule) -> (String, TopicRule) {
-    let key = if let Some(local) = topic.local() {
-        format!("{}/{}", local, topic.pattern())
-    } else {
-        topic.pattern().into()
-    };
-    (key, topic.clone())
-}
 
 /// Pump used to connect to either local broker or remote brokers (including the upstream edge device)
 /// It contains an mqtt client that connects to a local/remote broker
@@ -435,7 +236,7 @@ struct EgressShutdownHandle(Option<oneshot::Sender<()>>);
 impl EgressShutdownHandle {
     async fn shutdown(mut self) {
         if let Some(sender) = self.0.take() {
-            if let Err(e) = sender.send(()) {
+            if sender.send(()).is_err() {
                 error!("unable to request shutdown for egress.");
             }
         }
@@ -481,8 +282,6 @@ impl Egress {
         } = self;
 
         info!("starting egress publication processing");
-
-        let (_, shutdown) = oneshot::channel::<()>();
 
         let mut shutdown = shutdown_recv.fuse();
         let mut loader = loader.fuse();
